APPEALS-38907 Complete implementation of job minus va ops email

+++ app/jobs/hearings/download_transcription_file_job.rb
@@ -1,49 +1,59 @@
 # frozen_string_literal: true
 
-# TO-DO: confirm open-uri
 require "open-uri"
+
+# Downloads transcription file from Webex using temporary download link and uploads to S3
+# - Download link passed to this job from GetRecordingDetailsJob
+# - File type either audio (mp3), video (mp4), or vtt (transcript)
 
 class DownloadTranscriptionFileJob < CaseflowJob
   include Hearings::EnsureCurrentUserIsSet
 
-  # TO-DO: confirm priority
   queue_with_priority :low_priority
 
-  class TranscriptionFileNameError < StandardError; end
+  attr_reader :file_name
+
+  class FileNameError < StandardError; end
+  class FileDownloadError < StandardError; end
+  class FileUploadError < StandardError; end
   class HearingAssociationError < StandardError; end
 
-  retry_on(OpenURI::HTTPError, attempts: 10, wait: :exponentially_longer) do |job, exception|
-    Rails.logger.error("#{job.class.name} (#{job.job_id}) failed with error: #{exception}")
-
-    file_name = job.arguments.first[:file_name]
-    hearing_id = file_name.split("_")[1]
+  retry_on(FileDownloadError) do |job, exception|
     extra = {
       application: job.class.app_name.to_s,
-      hearing_id: hearing_id,
-      file_name: file_name,
+      hearing_id: job.hearing.id,
+      file_name: job.file_name,
       job_id: job.job_id
     }
-    Raven.capture_exception(exception, extra: extra)
+    log_error(exception, extra)
 
     # TO-DO
     # send_error_email_to_va_ops(exception)
   end
 
-  discard_on(TranscriptionFileNameError) do |job, exception|
-    Rails.logger.warn("Discarding #{job.class.name} (#{job.job_id}) failed with error: #{exception}")
+  retry_on(FileUploadError, wait: :exponentially_longer) do |job, exception|
+    File.delete(tmp_location) if File.exist?(tmp_location)
 
-    # TO-DO
-    # send_error_email_to_va_ops(exception)
+    extra = {
+      application: job.class.app_name.to_s,
+      hearing_id: job.hearing.id,
+      file_name: job.file_name,
+      job_id: job.job_id
+    }
+    log_error(exception, extra)
   end
 
-  VALID_FILE_TYPES = %w[mp4 vtt].freeze
+  discard_on(FileNameError)
+
+  VALID_FILE_TYPES = %w[mp3 mp4 vtt].freeze
 
   S3_BUCKET_NAMES = {
+    mp3: "transcript_audio",
     mp4: "transcript_audio",
     vtt: "transcript_raw"
   }.freeze
 
-  # Purpose: Downloads audio (mp4) or transcript (vtt) file from temporary download link provided by
+  # Purpose: Downloads audio (mp3), video (mp4), or transcript (vtt) file from temporary download link provided by
   #          GetRecordingDetailsJob
   #
   # Params: download_link - string, URI for temporary download link
@@ -54,20 +64,31 @@ class DownloadTranscriptionFileJob < CaseflowJob
     ensure_hearing_held
     @transcription_file = find_or_create_transcription_file
 
-    begin
-      download_file_to_tmp(download_link)
-      upload_file_to_s3
-      # queue_transform_job
-    ensure
-      clean_up_tmp_location
-    end
+    download_file_to_tmp(download_link)
+    upload_file_to_s3
+  end
+
+  # Purpose: Location of temporary file in tmp/transcription_files/<file_type> directory
+  #
+  # Note: Public method to provide access during job retry
+  #
+  # Returns: string, directory path
+  def tmp_location
+    File.join(Rails.root, "tmp", "transcription_files", file_type, file_name)
+  end
+
+  # Purpose: Hearing for which the transcription was created  #
+  #
+  # Note: Public method to provide access during job retry
+  #
+  # Returns: Hearing object
+  def hearing
+    @hearing ||= parse_hearing
   end
 
   private
 
-  attr_reader :file_name
-
-  # Purpose: Downloads audio (mp4) or transcript (vtt) file from temporary download link provided by
+  # Purpose: Downloads file from temporary download link provided by
   #          GetRecordingDetailsJob. Update file status of transcription file depending on download success/failure.
   #
   # Params: download_link - string, URI for temporary download link
@@ -75,38 +96,36 @@ class DownloadTranscriptionFileJob < CaseflowJob
   #
   # Returns: Updated @transcription_file
   def download_file_to_tmp(link)
+    return if File.exist?(tmp_location)
+
     begin
       URI.open(link) do |download|
         IO.copy_stream(download, tmp_location)
       end
       update_file_after_download(status: :success, date: Time.zone.now)
     rescue OpenURI::HTTPError => error
+      clean_up_tmp_location
       update_file_after_download(status: :failure)
-      raise error
+      raise FileDownloadError, error.message
     end
   end
 
-  # Purpose: Location of temporary file in tmp/transcription_files/<file_type> directory
-  #
-  # Returns: string, directory path
-  def tmp_location
-    File.join(Rails.root, "tmp", "transcription_files", file_type, file_name)
-  end
-
-  # Purpose: Downloads audio (mp4) or transcript (vtt) file from tmp directory to s3 location
+  # Purpose: Uploads file from tmp directory to s3 location
   #
   # Returns: Updated @transcription_file
   def upload_file_to_s3
     begin
       S3Service.store_file(s3_bucket, tmp_location, :filepath)
+      clean_up_tmp_location
       update_file_after_upload(status: :success, aws_link: s3_file_location, date: Time.zone.now)
-    rescue StandardError => error
+      log_info(s3_success_message)
+    rescue Aws::S3::Errors::ServiceError => error
       update_file_after_upload(status: :failure)
-      raise error
+      raise FileUploadError, error.message
     end
   end
 
-  # Purpose: Bucket name in s3 depending on if file type mp4 or vtt
+  # Purpose: Bucket name in s3 depending on if file type mp3, mp4, or vtt
   #
   # Returns: string, bucket name
   def s3_bucket
@@ -120,28 +139,21 @@ class DownloadTranscriptionFileJob < CaseflowJob
     s3_bucket + "/" + file_name
   end
 
-  # Purpose: Location of successfully uploaded file in s3
+  # Purpose: Builds on successful upload to S3
   #
-  # Returns: string, s3 filepath
-  def queue_transform_job
-    case file_type
-    when "mp4"
-      TransformMp4ToMp3Job.perform_later(s3_file_location)
-    when "vtt"
-      TransformVttToRtfJob.perform_later(s3_file_location)
-    end
+  # Returns: string
+  def s3_success_message
+    "TranscriptionFile (id: #{@transcription_file.id}) successfully uploaded to S3 location: '#{s3_file_location}'"
   end
 
   # Purpose: Removes temporary file (if it exists) from tmp folder after job success/fails
   #
   # Returns: integer value of 1 if file deleted, nil if file not found
   def clean_up_tmp_location
-    return unless file_name && File.exist?(tmp_location)
-
-    File.delete(tmp_location)
+    File.delete(tmp_location) if File.exist?(tmp_location)
   end
 
-  # Purpose: Either mp4 (audio) or vtt (transcript)
+  # Purpose: Either mp3 (audio), mp4 (video), or vtt (transcript)
   #
   # Returns: string, file type
   def file_type
@@ -154,14 +166,7 @@ class DownloadTranscriptionFileJob < CaseflowJob
   def file_type_invalid?
     return false if VALID_FILE_TYPES.include?(file_type)
 
-    fail TranscriptionFileNameError, "Invalid file type"
-  end
-
-  # Purpose: Hearing for which the transcription was created
-  #
-  # Returns: Hearing object
-  def hearing
-    @hearing ||= parse_hearing
+    fail FileNameError, "Invalid file type"
   end
 
   # Purpose: Parses hearing details from identifiers present in file name
@@ -174,7 +179,7 @@ class DownloadTranscriptionFileJob < CaseflowJob
       hearing_type = hearing_details.split("_")[2]
       hearing_type.constantize.find(hearing_id)
     rescue StandardError
-      raise TranscriptionFileNameError, "File name missing sufficient hearing identifiers"
+      raise FileNameError, "File name missing sufficient hearing identifiers"
     end
   end
 
@@ -224,7 +229,8 @@ class DownloadTranscriptionFileJob < CaseflowJob
   def update_file_after_download(status:, date: nil)
     @transcription_file.update!(
       file_status: Constants.TRANSCRIPTION_FILE_STATUSES.retrieval.send(status),
-      date_receipt_webex: date
+      date_receipt_webex: date,
+      updated_by_id: RequestStore[:current_user]
     )
   end
 
@@ -236,7 +242,13 @@ class DownloadTranscriptionFileJob < CaseflowJob
     @transcription_file.update!(
       file_status: Constants.TRANSCRIPTION_FILE_STATUSES.upload.send(status),
       aws_link: aws_link,
-      date_upload_aws: date
+      date_upload_aws: date,
+      updated_by_id: RequestStore[:current_user]
     )
+  end
+
+  # Purpose: Logging info messages to the console
+  def log_info(message)
+    Rails.logger.info(message)
   end
 end
